Add SEOユーティリティ関数を実装し、ライブラリ詳細ページのSEOメタデータを動的に生成
