+++ src/routes/user/libraries/[id]/+page.svelte
@@ -1,8 +1,16 @@
 <script lang="ts">
   import LibraryDetail from '$lib/components/LibraryDetail.svelte';
   import { createAppUrl, getLogoUrl } from '$lib/constants/app.js';
-  import { company_name, copy_count_update_failed } from '$lib/paraglide/messages.js';
+  import { copy_count_update_failed } from '$lib/paraglide/messages.js';
   import { getLocale } from '$lib/paraglide/runtime.js';
+  import {
+    addJsonLdToHead,
+    generateJsonLd,
+    generateKeywords,
+    generateSeoDescription,
+    generateSeoTitle,
+    removeJsonLdFromHead,
+  } from '$lib/utils/seo.js';
   import { onMount } from 'svelte';
   import type { PageData } from './$types.js';
 
@@ -19,111 +27,20 @@
   // 現在のロケールを取得
   const currentLocale = getLocale();
 
-  // SEO用のtitleとdescriptionを動的に取得
-  const getSeoTitle = () => {
-    if (!librarySummary) return `${library.name} - AppsScriptHub`;
-
-    if (currentLocale === 'ja' && librarySummary.seoTitleJa) {
-      return librarySummary.seoTitleJa;
-    } else if (currentLocale === 'en' && librarySummary.seoTitleEn) {
-      return librarySummary.seoTitleEn;
-    }
-
-    // フォールバック
-    return `${library.name} - Apps Script Hub`;
-  };
-
-  const getSeoDescription = () => {
-    if (!librarySummary) return library.description;
-
-    if (currentLocale === 'ja' && librarySummary.seoDescriptionJa) {
-      return librarySummary.seoDescriptionJa;
-    } else if (currentLocale === 'en' && librarySummary.seoDescriptionEn) {
-      return librarySummary.seoDescriptionEn;
-    }
-
-    // フォールバック
-    return library.description;
-  };
-
-  // 重複のないキーワードを生成
-  const generateKeywords = () => {
-    const baseKeywords = ['JavaScript', 'Google Apps Script'];
-    let additionalKeywords: string[] = [];
-
-    if (librarySummary) {
-      if (currentLocale === 'ja' && librarySummary.tagsJa) {
-        additionalKeywords = librarySummary.tagsJa;
-      } else if (currentLocale === 'en' && librarySummary.tagsEn) {
-        additionalKeywords = librarySummary.tagsEn;
-      }
-    }
-
-    // 重複を除去してキーワードを結合
-    const allKeywords = [...baseKeywords, ...additionalKeywords];
-    const uniqueKeywords = [...new Set(allKeywords)];
-    return uniqueKeywords.join(', ');
-  };
-
-  // JSON-LD構造化データを動的に生成
-  const generateJsonLd = () => {
-    const pageUrl = createAppUrl(`/user/libraries/${library.id}`);
-
-    // Software Source Code Schema
-    const jsonLd = {
-      '@context': 'https://schema.org',
-      '@type': 'SoftwareSourceCode',
-      name: getSeoTitle(),
-      description: getSeoDescription(),
-      url: pageUrl,
-      programmingLanguage: 'JavaScript',
-      runtimePlatform: 'Google Apps Script',
-      codeRepository: library.repositoryUrl,
-      author: {
-        '@type': 'Person',
-        name: library.authorName,
-        url: library.authorUrl,
-      },
-      license: library.licenseUrl,
-      dateModified: library.lastCommitAt.toISOString(),
-      ...(library.starCount && {
-        interactionStatistic: {
-          '@type': 'InteractionCounter',
-          interactionType: 'https://schema.org/LikeAction',
-          userInteractionCount: library.starCount,
-        },
-      }),
-      keywords: generateKeywords(),
-      publisher: {
-        '@type': 'Organization',
-        name: company_name(),
-        url: 'https://wywy.jp/',
-      },
-    };
-
-    return jsonLd;
-  };
+  // SEO関連のデータを取得
+  const getSeoTitle = () => generateSeoTitle(library, librarySummary, currentLocale);
+  const getSeoDescription = () => generateSeoDescription(library, librarySummary, currentLocale);
+  const getSeoKeywords = () => generateKeywords(librarySummary, currentLocale);
+  const getJsonLd = () => generateJsonLd(library, librarySummary, currentLocale);
 
   // コンポーネントマウント時にJSON-LDを動的に追加
   onMount(() => {
-    // 既存のJSON-LDスクリプトがあれば削除
-    const existingScript = document.querySelector('script[type="application/ld+json"]');
-    if (existingScript) {
-      existingScript.remove();
-    }
-
-    // 新しいJSON-LDスクリプトを作成
-    const script = document.createElement('script');
-    script.type = 'application/ld+json';
-    script.textContent = JSON.stringify(generateJsonLd());
-    document.head.appendChild(script);
+    // JSON-LDを追加
+    addJsonLdToHead(getJsonLd());
 
     // クリーンアップ関数を返す
     return () => {
-      const scriptToRemove = document.querySelector('script[type="application/ld+json"]');
-      if (scriptToRemove) {
-        scriptToRemove.remove();
-      }
+      removeJsonLdFromHead();
     };
   });
 
@@ -203,7 +120,7 @@
 <svelte:head>
   <title>{getSeoTitle()}</title>
   <meta name="description" content={getSeoDescription()} />
-  <meta name="keywords" content={generateKeywords()} />
+  <meta name="keywords" content={getSeoKeywords()} />
   <meta name="author" content={library.authorName} />
 
   <!-- Open Graph tags -->
